feat(inventario): actualizar escala progresiva de puntos por conteo diario

+++ app/Models/InvoiceCount.php
@@ -40,7 +40,7 @@ class InvoiceCount extends Model
         'counted_quantity' => 'float',
         'system_quantity' => 'float',
         'discrepancy' => 'float',
-        'points_earned' => 'integer',
+        'points_earned' => 'float',
         'penalty_points' => 'integer',
         'created_at' => 'datetime',
         'updated_at' => 'datetime',
